+++ sdk/test-utils/perfstress/src/utils.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import { IncomingMessage, RequestOptions, request } from "http";
-import { TestProxyHttpClient, TestProxyHttpClientV1 } from "./testProxyHttpClient";
 
 /**
  * Returns the environment variable, throws an error if not defined.
@@ -41,28 +40,6 @@ export async function makeRequest(
 
     req.end();
   });
-}
-
-const _cachedProxyClients: {
-  v1: TestProxyHttpClientV1 | undefined;
-  v2: TestProxyHttpClient | undefined;
-} = {
-  v1: undefined,
-  v2: undefined
-};
-
-export function getHttpClientV1(url: string): TestProxyHttpClientV1 {
-  if (!_cachedProxyClients.v1) {
-    _cachedProxyClients.v1 = new TestProxyHttpClientV1(url);
-  }
-  return _cachedProxyClients.v1;
-}
-
-export function getHttpClient(url: string): TestProxyHttpClient {
-  if (!_cachedProxyClients.v2) {
-    _cachedProxyClients.v2 = new TestProxyHttpClient(url);
-  }
-  return _cachedProxyClients.v2;
 }
 
 /**